requirements.txt
aiohttp~=3.8.4
requests~=2.29.0
beautifulsoup4~=4.12.2
nonebot2>=2.0.0b4
aiohttp
Pillow
icalendar
emoji
aiohttp_socks
fuzzywuzzy